VariantType::extract_slice() now takes a signature argument
This is not needed for simple types (which we've been dealing with so far)
which have static signatures but we'll need this for complex types (e.g
arrays, structures and dictionaries) where the signature will be dynamic
and possibly complex.

To keep things simple for simple types, we add a `SimpleVariantType` trait
that only has a helper function: `extract_slice_simple` which simply calls
the `extract_slice` with the constant signature for the type.