VariantType::decode() now takes a signature argument
This is not needed for simple types (which we've been dealing with so far)
which have static signatures but we'll need this for complex types (e.g
arrays, structures and dictionaries) where the signature will be dynamic
and possibly complex.

To keep things simple for simple types, we add `decode_simple` helper
method to `SimpleVariantType` trait that simply calls the `decode` method
with the constant signature for the type.