Add bytes_before arg to VariantType::decode() & extract_slice()
Similar to VariantType::encode(), the decoding & parsing methods now
decode/parse padding themselves and hence need to be informed of the
number of packed bytes before them.

The reason is that the individual data items in a stream/message are not
aligned in relation to the immediate parent container but the entire
stream/message.