New enum based API
Variant is now an enum and hence carry its value in raw/decoded format.
This means a lot of API had to be changed and new API had to be
introduced. This includes:

* Removal of generics from the VariantType implementers. They themselves
  now keep their children as Variants but provide generics-based API for
  convenience and further type-safety.
* New types:
  * Array is now an actual type that can be transformed from/to a
    Vec<T: VariantType>.
  * Dict is now an actual type that can be transformed from/to a
    Hashmap<K: SimpleVariantType, V: VariantType>.
  * MessageFields: Just a wrapper around Vec<MessageField>.
* MessageField no longer implements VariantType but rather provides
  Deref<Target=Structure> and DerefMut for the underlying Structure.

Some other changes included too probably but 2020 Zeeshan, you know I've
been working on this for months now and I just want to get this finally
done so I'm sorry for handling over a large commit with multiple changes.
It's really hard anyway to split a patch that changes the API on such a
fundamental level. So sorry, but not very sorry.