Take array's first element padding into account when encoding/decoding
This fixes the main issue I've been fighting for the last 2 months now at
least. I thought I had misunderstood Variant's encoding but turns out now
that that wasn't the issue after all.

The real issue is that the size of the array that's encoded in the array
itself, does not include the padding of the first elmeent.