Drop redundant AlignmentKind type & its use
Turns out that my initial assumption of Variant's value needing encoding
was correct after all so there is no need to pass that around and adapt it
since alignement is always done, regardless of the parent/context.

We still keep the EncodingFormat part of the context and pass it around as
we'll need that when adding GVariant format support.