Serializer now requires VariantValue trait
This is not ideal at all but we need the signature of the whole value,
cause otherwise we've no way of adding the first element padding in an
array if it's empty. So blame D-Bus!

However, I'll provide blanket implementations for VariantValue (we already
do that) and a macro to implement VariantValue on custom types.