Replace DictEntry by Dict
DictEntry can't have VariantValue implemented for it so it can only be
used with Variant and there is no point in supporting standalone DictEntry
only in Variant. We already support HashMap so a Dict will be the Variant
equivalent for that. DictEntry is now an internal detail of Dict.

Like in 1.0 API, we provide conversions between HashMap and Dict.