Custom traits for Variant conversions
While it'd have been awesome if we could make use of `*From` and `*Into`
traits from `std` crate, we can not override the blanket implementations
of Variant to Variant conversions and we need to flatten/deflatten it to
keep the API consistent with Variant conversions for other types.

Hence introducing custom traits, `FromVariant` and `IntoVariant`.