zv: remove the owned version of Value::string
We only need borrowed kind, avoiding mixing borrow/owned version in
Value. There should be a owned-Value version instead.

Note that this removes the ability to convert a char to a string. We
could argue that char isn't strictly a string either though (a char is
a 'Unicode scalar value', which uses a different representation than
what string uses)

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>