zv: replace FromValue with TryFrom
FromValue is a custom trait that isn't really convenient. In general,
it does what TryFrom is supposed to do, plus it downcast. Instead,
implement the idiomatic TryFrom for the various Value kind, and use
Value::downcast_ref() when explicit downcast is expected.

This is quite intrusive, so I don't have a good strategy to reduce the
churn. If necessary, we can probably introduce the TryFrom first, and
then remove ValueFrom, but that's not really helping imho.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>