zb: add Connection::next_message()
Extract out the code to get the next incoming message, and refactor
the inner loop in call_method().

Note that instead of ignoring/println a parsing error, it will now
return the error immediately to the caller, who can take the decision
to continue, if it needs to handle such scenario. It keeps ignoring
other well-formed messages though, for now.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>