zb: require Any for Interface
Since we already require that dyn Interface must be 'static to be in the
ObjectServer, it's implicit that Any is already provided (it's provided
for all T: 'static). Make this explicit on the Interface trait, so that
we can further extend it with a function to return dyn Any of self.

Also add an internal mechanism to downcast a `dyn Interface` to a
concrete type. We'll make use of this in a following commit.

Co-authors:

Zeeshan Ali <zeeshanak@gnome.org>
Sebastian Dröge <sebastian@centricular.com>