zv: Test case for ostree data deserialization
While this involves keeping a 7mb file in the repo, the test case is
extremely good as it deserializes a big chunch of very complex set of
data, that uncovered some interesting issues in the code. Nevertheless
it's behind a new feature and data is kept in the root of the repository
so that it's never included in the crate distribution.

Later, we should replace this with some fuzzing-based solution, I think.