zb: introduce RawConnection
This introduces a RawConnection abstraction, which handles only
serialization and deserialization of D-Bus messages over a socket,
in a way compatible with non-blocking sockets (doing internal
buffering to handle partial read/writes).

Also introduce the Socket trait, as a general abstraction over
the transport for the connection. This is to ease future integration
with runtime-aware socket types, for async compatibility.