zb: support ServerHandshake on *BSD
SO_PEERCRED is a Linux-specific option (also available on OpenBSD,
but the nix crate only exposes it on Linux).

FreeBSD has LOCAL_PEERCRED, but it's also not exposed by nix.

However, we only need the uid here (not the pid), so getpeereid(), which
is exposed for all BSD-like systems in the nix crate, is the way to go.

In addition, allow receiving SCM_CREDS on FreeBSD/DragonFly,
as clients do send it to allow the server to get the pid.