zv: Allow disabling `GVariant` support with a feature
Since GVariant supports `Option<T>` (Maybe type) and we can't make that to
anything in D-Bus, people end up trying to send/revceive `Option<T>` on
D-Bus (through `zbus`) and only get an error at runtime.

Let's add a feature, `gvariant` so that gvariant-specific API/impl can
be disabled by `zbus` to avoid this situation. There may be some other
uses of this too.

We should also add a `dbus` feature but that'd be harder to implement
than `gvariant` and we currently have no use case for that so not adding
it yet.