zb: make Connection wrap a dyn Socket
This will allow various Socket implementations to be used as Connection.

I don't see much need to make this generic over S: Socket, and it would
be a big API breakage. Another potential option is to have an enum for
known Socket types, but this will likely not have any performance impact
and just make the code more complicated and bloated.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>