zb: Drop generics completely from public azync::Connection
This commit turns the generics of azync::Connection into a private
detail. This means users don't have to worry about which transport type
is being used underneath.

This also implies that users can't just plug-in a transport in a
connection. We may provide a way to do that in the future using trait
objects (which is what we use internally now) or encourage people to add
transports directly into zbus itself.