zb: Separate sockets for read & write in azync::Connection
At the moment, this only make azync::Connection more consistent with
Connection but doesn't give us any advantage but it will when we move to
interior mutability in a following commit. Then it'll help us avoid
deadlock when user reads and writes to the same connection at the same
time.