ci: use fdo ci-template
Main feature is to install all the packages and tools we need
to the container image, saving us to download and build them at each
run.

I used the same strategy as gstreamer-rs's ci: rely solely on container
for caching instead of using gitlab's cache.
This seems to be faster now that all tools are in the image as saving
and restoring the cargo and target cache was taking a while.
It also reduces the data transfers between runners which are the most
expensive part of the bill for the ci infrastructure.

The image needs to be manually updated when a new version of Rust is
released. This is easily done by bumping FDO_DISTRIBUTION_TAG.
Doing so also reduce the chance of CI breaking behind our back because
of Rust or tools updates.