zb: Make incoming msg queue, leaky
Instead of dropping all new messages after the queue is full, drop the
oldest msg in the queue to make room for the new one, instead. With the
queue size as large as 64 by default, if the message hasn't yet been
picked up by anyone, it likely never will be and the new messages might
be of more interest.