zv: Allow converting Infallible to Error
This enables users (zbus mainly) to satisfy `TryFrom` requirements with
`From` (and `TriInto` with `Into`) implementations.

Unfortunately, this is an API break since error type inference in
certain cases won't work and user will have to explicitly specify error
types. See changes to `utils.rs` file in this commit for an example.