zb: Interior mutability for FDs in a Message
Users shouldn't need to have mutable ref to incoming messages and
disowning FDs was the only real reason a user would need it so let's
just use interior mutability for FDs so users rarely need mutable ref to
a message.

This change becomes more important later on when we switch to
`Arc<Message>` for incoming messages.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>