zb: Remove a block_on call from a Drop::drop impl
While it'd be great to be able to cancel the message receiver task and
wait for that before a connection is dropped, calling `block_on` from
`Drop::drop` impls is problematic, keeping in mind the very common case
of Connection being dropped in async contexts.

Let's rely on task being cancelled when it's dropped.