zb: Add ObjectServer::request_name
Since typically a well-known name is registered by services, it makes sense
to provide the most convenient way to register it. Theoretically this
should be a method of `Connection` but there is a very real a danger of
users registering their name before creating the `ObjectServer` instance
and as a result, missing out the first method call (e.g in case of D-Bus
autolaunch).

The method uses the builder pattern as typically this will be used when
instantiating `ObjectServer`.

All the names registered through this new method are automatically
released on `ObjectServer` dropping but a `release_name` method is
provided for explicitly releasing the names.

Fixes https://gitlab.freedesktop.org/dbus/zbus/-/issues/114.