zb: azync::Connection implements Stream directly
instead of through MessageStream, which is dropped as part of this
change. While a mutable ref is needed to operate on streams, we don't
need to switch to `&mut self` in other methods since one can just clone
the Connection to get a mutable ref to it. i-e All `stream()` calls can
just be changed to `clone()` calls.