zb,zm: Make use of InterfaceName type
This is to ensure that this string type is checked for conformance. We
should do that anyway but the D-Bus specification also mandates
implementations to verify everything to and from the bus.

While generic implementations and `TryFrom` implementations ensure that
impact to users is minimum with this change, unfortunately they do now
have to specify type when passing a `None` value. This howeover is a
well-known pain-point in the Rust world and nothing specific to zbus.

Our currently solution is close to what will come in the future with `!`
types: https://github.com/rust-lang/rust/issues/35121 .