zb: Change ProxyProperties::values to sync Mutex
The cached properties do not need to be protected by an async mutex
because all updates to the cache are in response to a message we already
have without blocking.  This is required to handle reads from the cache
without blocking, and does not impact the normal get functions that
already do an async Get call when the cache does not contain the value.