zb: Add helper trait for zbus_macros
This trait allows macros to refer to the `Ok` and `Err` types of a
Result that is behind a type alias.  This is currently required
because the macros for properties expect a Result return value, but the
macro-generated `receive_` functions need to refer to the actual type
without the associated error.