zb: Don't require proxy destination to be known
Let's allow users to create proxies for inexistant destinations. This
makes sense anyway but many services rely on autostart and keep
themselves deactivated when not being used.

Since receiving a signal from the service means it's up, we now only try
to resolve the name on signal reception but if it fails, the signal just
doesn't match but we don't treat it as an error.