zb: Disable overflow on Connection's broadcast channel
With overflow, there are scenarios where messages could be lost. This
would be not that huge an issue if we provided a way to detect when this
happens but we don't (and can't easily). Even if we provided such an
API, not every message receiver would be able to recover from this loss
of data. Hence disabling overflow mode.

However, OTOH not overflowing can easily hang the Connection if one of
the streams/receiver doesn't actually receive. We will address this
issue through various changes in our API in the following commits.

Fixes #178.