zb: Bring back azync::MessageStream
Now that we've disabled overflow mode on the receiving side of the
Connection's internal broadcast channel of inbound messages and as a
result, require Connection to be continously polled as a stream, let's
provide users a way to keep a form of connection around w/o having to poll
it all the time. We do so by re-introducing our old friend MessageStream.

This however is only a solution for low-level API. The high-level API
(especially the sync one) currently uses the stream behind the scene but
depends on the user to keep polling the stream indirectly. We'll resolve
this problem in the following commits by moving the polling into a task
run by the Connection executor and thus unsuring the internal streams
continously get polled.

I initialially (in my WIP branch) kept the MessageStream keep a
Connection so that stream can be converted back into a Connection but
that leads us into more cyclic Arc references than we already have. Also
it doesn't really give user much. If they want to keep both, they can
just keep both a Connection and a Stream around and that's no issue.