zb,zm,book: Proxy signal handlers now called automatically
We now run a task in our executor that listens for signals and calls the
appropriate handlers for the user. This simplifies the API and removes a
footgun: user not reading the docs carefully and not calling `next_signal`
method and wondering why their signal handlers never get called.

Fixes #188.