zb,book: Don't require Result return from proxy signal handlers
Now that user-provided signal handlers are called implicitly by us and
we've no way of forwarding the errors from the handlers back to the
user (unless they use `Proxy::handle_signal` which won't be the typical
use case and we might actually want to drop that API all together), its
best we don't ask for result to be returned by the callbacks.

This is also consistent with property change handlers.