zb,book: Move name registration to Connection
Move name registration from ObjectServer to Connection. Names are
registered against connections anyway but this API was added to
ObjectServer as typically names are requested by services. Now we want
to create a 1-1 association between Connection and ObjectServer in a
following commit, and this API on ObjectServer gets in the way of that
change.

This temporarily re-introduces the issue of the initial method call
being lost on autolaunch of the D-Bus service (see #68), which was one
of the reasons this API was added on ObjectServer rather than
Connection. This issue will be resolved in a following commit that
establishes the 1-1 association between Connection and ObjectServer.