zb,zm,book: Async ObjectServer & dbus_interface generates azync::Interface
This commit:

* adds azync::ObjectServer that's well, async version of ObjectServer.
  It uses azync::Interface for that reason.
* turns crate::ObjectServer into a thin wrapper around
  azync::ObjectServer.
* modifies dbus_interface to generate azync::Interface implementation.

I really wanted to split this change into multiple commits with smaller
steps but the dbus_interface and ObjectServer code is so inter-dependent
that that proved to be super difficult. It'd have required a lot more of
my time, which I don't have in sufficient amount. It'd have also
required writing a lot of code, only to be thrown away soon after. Hence
this one big commit with multiple logical changes in. May the git gods
forgive me!

Fixes #37.