zb, zm: Add DispatchResult
This dispatches the call before returning the Future.  This will
generally cause the allocated Future objects to be smaller and avoids
needing to take a write lock in order to handle "method not found".  If
an interface contains no `&mut self` methods, this means that no write
locks will ever be taken.