zb: Process MethodCall messages in parallel
Without this, the MessageStream used by the ObjectServer's task will not
be polled until a call has completed.  This could cause a deadlock if
the async method makes calls using dbus and the queue fills up before
the response to one of its calls is added.

Interfaces already use locking to ensure they only handle one mutating
call at a time, and dbus clients that depend on call ordering will wait
for the first call to return before making the second.