zb: Add Connection::new_scope
A scope is a set of connect_* callbacks that are well-ordered with
others in the same scope (one message is fully handled before starting
the next).  This allows creation of more than one scope for a given
Connection object, which is useful if some callbacks are slow to run.
See the documentation of that method for full detials.