zb: Add Connection::call_method_raw
This exposes an OrderedStream for method calls, which is needed when the
reply must be processed in order with other messages (such as cache
update signals).

This structure is mostly equivalent to `msg_stream.filter(...).take(1)`,
but with a type that can be written instead of one containing a closure.