zb: Add ConnectionBuilder::serve_at
Add API that allows users to register their interfaces before the
connection is established and therefore before object server is started.
This not only removes the need for p2p connections to have to start the
object server explicitly (`ObjectServer::start_dispatch` will be dropped
in the next commit) but will also provide an easy way to avoid footgun
resulting from name registration before interface registration in a
following commit.