zb,zm: Abstract away property invalidation
Property invalidation is the same as property update, except the new
values are not communicated as part of the signal. This is nothing a
user needs to know about (if they do, they can use low-level API) so we
now have `PropertyStream` yield a new type, `PropertyChanged` rather than
the strange `Option<Option<T>>`.

If users are interested in the new value, they can use
`PropertyChanged::get`, which calls the `Get` call to retreive the new
value in case of invalidated properties.