zb: Add MessageSequence
This exposes the relative ordering of two messages received on the same
connection.  While this could be done using timestamps, there are a
number of issues that that causes (Instant is only non-decreasing and we
need increasing, SystemTime can be set, and the timestamp is not what
the user most likely wants - which is the time it was seen on the bus).
Because we don't actually expose it as a number, we could switch later.