zb: Make SignalStream independent of callbacks
This reworks SignalStream to be based directly on the broadcast channel,
basically reverting d2cd55565a and switching to an alternative
mechanism.  This avoids depending on the message dispatch task and makes
it easier to later implement OrderedStream for SignalStream.

The GetNameOwner query is made specific to the signal stream because the
shared value may not match the actual value at the time of the signal
delivery.  This could cause incorrect message drops in either case: if
the SignalStream is behind (because its task was run after the
cache/callback task) then it will drop a signal sent just before the
sender releases its name.  If the SignalStream is ahead, then it could
miss the first signal that was sent just after a name was acquired.

In the future it may be possible to re-enable caching of GetNameOwner,
but this requires more extensive changes.