zb: impl OrderedStream for SignalStream
This allows taking two different SignalStream objects associated with
the same connection and producing a composite stream of messages whose
order is the same as that seen on the bus.  This is useful if you need
to handle more than one signal and would like to ensure you see causes
before effects.