zm,docs: Don't run dbus_proxy example as test case
Only building is enough. This proxy testcase doesn't setup the
corresponding interface so the calls etc fail anyway. On top of that and
we're about to drop the callback based signal API and it'd be hard to
modify the code to nice fail and still be a nice little example that it
is. We've other tests that cover all this functionality.