zv: Reduce size of Str from 4 to 3 pointers
Because we do not expose an API to modify the contents of a Str, storing
extra capacity wastes space - both in the allocation and in needing an
extra word to track it.  Switching from String to Box<str> avoids this.

This also adds a third variant for statically owned strings that can be
used to avoid allocations in to_owned.  This is a fairly common
situation when strings are names of objects on the bus.