zb,book: Introduce InterfaceRef
This removes get_interface/_mut in favor of a single interface(path)
function that returns an InterfaceRef, which in turn has get/_mut
methods to access the InterfaceDeref/Mut instances. More particularly,
this ensure that the lifetime of InterfaceDeref/Mut is not tied to the
originating Node, meaning that the InterfaceDeref can be used without
locking the Node or ObjectServer.