zm,zb: Correctly convert zbus::Error::Method to DBusError impls
This brings back functionality dropped in 986de17488bf6d, where started
to convert zbus::Error to always convert to the fallthrough `zbus_error`
variant. This would mean that proxies using fdo::Error (which is a
DBusError) will not actually be able to make use of this error. The same
would go for other DBusError impls as well. This is of course not nice.

So this commit fixes that by converting to the appropriate variant, if
any.