zm: Fix DBusError docs
Remove two lies:

* We can't actually support `&'static str` type for description fields.

* We don't support >1 fields in variants.

Both of these are because of the `From<zbus::Error>` impl so when that's
not requested (by lack of the special `zbus_error` variant), we could
support both but right now I'm too lazy to fix that and explain that in
the docs. We can add that in the future.