zb: Run ConnectionBuilder::build futures w/o executor
Currently, if any name is requested through ConnectionBuilder::name
method and internal executor is disabled, we end up in a hang since the
external runtime only kicks in after we've returned from build call.
Let's ensure that all the futures in the ConnectionBuilder::build are
run even if there is no internal runtime ticking thread.