zx: Do not generate unsized `[T]` types
Not sure if there's a better way to improve the logic here, but this
seems to be correct. I don't think the `[T]` binding would be valid
anywhere since this doesn't seem to be used anywhere an unsized type
would be valid. So this can't break anything.

Fixes https://gitlab.freedesktop.org/dbus/zbus/-/issues/192.