zd: Type macro to allow specifying hardcoded signature
There are cases where you know better than the macro and manually
implementing the `Type` macro is a bit of a work so this can be handy.
This can be used for all kinds of structs and enums that the macro can't
(currently) support.

Moreover, we can use this also for the "struct as dict" case and
deprecate the `DictType` macro in a following commit.