zb: Connection::request_name no longer starts the ObjectServer
Otherwise it becomes impossible to implement a low-level service
implementation (which is included as an example in the book).

As for avoiding issue#68, we've to leave that to the user. We do provide
convenient API to easily avoid it and we also document the pitfall in
the [book] (with clear instruction on how to avoid it).

Fixes #260.

[book] https://dbus.pages.freedesktop.org/zbus/server.html#-service-activation-pitfalls