zv: Correctly ser/de non-unit enums
It's very wrong to encode data (the enum variant index) w/o it being
represented in the signature. Moreover, variants of the same enum can't
have different signatures and if this would be used in practice, it'd
blow up in many ways.

We fix these issues by now assuming all variants of an enum to always
have the same number and types of arguments (and hence the same
signature) and encoding all non-unit enums as a structure with variant
index as the first field and variant's field as the second one. So:

```rust
// This gets signature `(us)`.
enum NewType {
   Var1(String),
   Var2(String),
}

// This gets signature `(u(yt))`.
enum Struct {
    Tuple(u8, u64),
    Struct { y: u8, t: u64 },
}
```