zv: More robust check for non-unit enum deserialization
We were checking for a `)` to check if we just deserialized a non-unit
enum but this would also be true if we just deserialized a unit enum and
we're at the end of a containing structure (i-e a unit enum at the end of
a structure). This meant that we prematurely parse the `)` and if this
happens to be end of a message, we get a crash when the structure
deserializer tries to parse the signature still (as it is still
expecting at least a `)`).

Fixes #265.