zb,test: Don't expect property caching to be slow
The property stream creation triggers the (lazy) property caching in a
separate task in the Proxy implementation. The assertion immediately
after the stream creation was assuming the caching to be slow and hence
the property to be unset. This can not be relied upon so let's not make
this assumption.

The downside of this change is that the test doesn't always get to
actually test the fix from !460 but that's much better than having a
test case that can randomly fail on us.