zb: Enable tokio integration if `tokio` dep feature is enabled
This solves the issue of tokio integration not getting enabled if any of
the deps in a project doesn't disable async-io, even if you do. However,
this means things can break easily for async-io users if tokio
integration is enabled by another crate. i-e we end up with the same
problem but now the victims are the async-io users.