zv: Make it possible to serialize struct as ARRAY
Since serde serializes Rust arrays as structs, this can sometimes create
problems for us. For example, `std::net::IpAddr` enum can not have a
static signature and hence can't implement `Type`. We could implement
`DynamicType` for it but that's not a simple type to deal with. So let's
just allow serializing struct as an ARRAY if given the signature of a
ARRAY.

It should be noted that our deserializer can already do the opposite so
no changes needed there.