zb: Add Socket::send_zero_byte() for freebsd & dragonfly
This should be something that Socket takes care of as it depends on the
socket implementation type. This will allow us to remove some platform
specific code from handshake code and in a following commit also allow
users to be able to plug-in Socket impls that can handle this for us.