zb: Require data with AUTH ANONYMOUS
The relationship between AUTH ANONYMOUS and any data is not explained in
the D-Bus spec itself, and if the corresponding SASL RFC says it, no one
has bothered to read it (that does include this commit's author). In
particular, different D-Bus p2p server & client implementations handle
this a bit differently:

- GDBus's server does not accept a separate DATA message at all, and it
  will fail if one is given. Initial data may be passed, however, but it
  will be ignored. If no initial data is passed, the WaitingForBegin
  state is entered.
- GDBus's client always passes some initial data with AUTH ANONYMOUS.
- sd-bus's server *must* be given either initial data or a separate DATA
  message.
- sd-bus's client always passes some data in a separate DATA message,
  but not with the initial message.

In order to try to handle all of these, AUTH ANONYMOUS is treated like
AUTH EXTERNAL: data is required, whether it be passed via initial data
or a separate DATA message. This matches the sd-bus server's behavior
but also should be compatible with GDBus clients.