Generalize Sink impl of Connection to T: Into<Arc<Message>>
Not only this makes the API more consistent with that of MessageStream,
this also means users can simply connect a MessageStream to Connection
(as Sink) and have the sink consume the stream.

We also manage to not break the API because `From<T> for Arc<T>` is
implemented by std.