zv: Drop a wrong assertion about empty framing offsets
When each element of a sequence/struct is serialized, the serializer has
no idea about the possible existance of any following elements/fields so
it has to push the framing offset to the list. So when the container's
serialization ends, we don't want to treat existance of a framing offset
in the list as an error if the serialized container is of 0 length.