zb: Drop explicit channel close calls from Proxy caching code
Turns out that our workaround to closing the tokio's Sender via Receiver
ends up in a hang so we'll be dropping Sender::close in a following
commit. So this change is needed.

We achieve this by splitting the code that initializes the properties
cache and one that keeps it up2date.

Fixes #279.