zb: ObjectServer to only serve methods destined to its connection
While in most cases, you'd want ObjectServer to take over all incoming
method calls on its connection but in certain cases, you don't. For
example, if you'd build a bus implementation on top of zbus, you'll
likely want to make use of ObjectServer for the Freedesktop standard
interfaces required from a bus, while being able to receive messages on
the same connections destined for other peers to send them off to the
right connection. If ObjectServer would serve all messages, it'll just
return an error for all method calls destined for other peers.

I'm sure there are other (not so niche) use cases for this.