zm: add new attribute 'assume_defaults' to dbus_proxy macro
This is the initial step towards reworking `dbus_proxy` approach
to defaults.
It introduces a new `assume_defaults` attribute (defaulting to
`true`), which reflects the status quo: currently, all generated
proxies assume (and auto-generate) some default values for path
and service. In some cases, the auto-generated values could be
wrong and misleading.
The new attribute allows to selectively disable the implicit
default.