zb: Enable `tracing` feature of tokio
We'll need this in  following commit to enable experimental API to give
names to tasks so tokio-console can show our tasks by name and users/us
will have a much better luck reading the console output.

`tokio_unstable` cargo flag will need to be specified though but we
leave that to the users.