zb: Proxy::call_with_flags::flags shouldn't be generic
There is not much point in keeping it generic IMO. It only saves the
caller a `.into()` call if they want to specify only one flag. Given
this will be a very niche use-case (especially since we've more specific
`call_noreply` method too, I think the price of generics isn't worth it
in this specific case.