zb: Simplify SignalStream impls
Let's rely on MessageStream to handle the ordered_stream::OrderedStream
impl for us.

This commit mostly revert changes from d365b7f. That commit was written
and merged when we had internal task to call user callbacks, which was a
major reason for those changes I believe (the commit log seems to concur).
We removed that not too long after d365b7f but seems we kept these
changes because they weren't really harming anyone.