Revert "zb: Don't match signal sender if proxy destination unknown"
This reverts commit 845fd2463d6a50f26948f41cdf76d3a861dd24db.

Now that we resolve the destination before user starts polling the
stream, we can require sender of the signal to match the destination
unique name.

There is a chance we may miss an emission of the target signal if it's
sent before destination registers its name but that's not our fault and
not our problem to worry about IMO. Services should not do such a thing.