zv: Implement `io_lifetimes::AsFd` and `From` conversion with `OwnedFd`
This provides at least safe interop with standard IO safety types, to
the extent that's practical without breaking changes
(https://gitlab.freedesktop.org/dbus/zbus/-/issues/286).

For `Fd` this is more problematic since that lacks a lifetime and cannot
guarantee it contains a valid fd.

In recent Rust version, this is a re-export of the std type. When zbus
depends on a Rust version that has this, the dependency can be dropped.